docker-for-flask-bak/requirements.txt
celery==5.3.6
flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
redis==5.0.1
PyYAML==6.0.1